add Hanime1Parser and support for hanime1.me

+++ CSWebDriverClient/Models/Requests/PressButtonOnPageRequest.cs
@@ -7,9 +7,9 @@ public class PressButtonOnPageRequest
     [JsonPropertyName("button_xpath")]
     public string ButtonXPath { get; set; }
     [JsonPropertyName("timeout")]
-    public float Timeout { get; set; }
+    public double Timeout { get; set; }
 
-    public PressButtonOnPageRequest(string buttonXPath, float timeout = 5)
+    public PressButtonOnPageRequest(string buttonXPath, double timeout = 5)
     {
         ButtonXPath = buttonXPath;
         Timeout = timeout;
